backend/flask/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
